+++ mods/unblur.user.js
@@ -1,0 +1,15 @@
+function unblurInit (toggle) {
+
+    const unblurCSS = `
+    .thumb-subject, .image-filler {
+        filter: none !important;
+    }
+    `;
+
+    if (toggle) {
+        safeGM("removeStyle", 'unblurred');
+        safeGM("addStyle", unblurCSS, 'unblurred');
+    } else {
+        safeGM("removeStyle", 'unblurred');
+    }
+}